Finished the functionalities of the Detail page. Nullable response data type is still causing error (to be fixed in the next push).

+++ core/src/main/java/com/rivibi/mooviku/core/data/remote/response/GetDetailResponse.kt
@@ -20,7 +20,7 @@ data class GetDetailResponse(
 	val backdropPath: String,
 
 	@field:SerializedName("revenue")
-	val revenue: Int,
+	val revenue: Long,
 
 	@field:SerializedName("genres")
 	val genres: List<GenresItem>,
@@ -38,7 +38,7 @@ data class GetDetailResponse(
 	val voteCount: Int,
 
 	@field:SerializedName("budget")
-	val budget: Int,
+	val budget: Long,
 
 	@field:SerializedName("overview")
 	val overview: String,
